Update About_Push_Pull_and_Fetch
Here are the questions I embedded in the doc in the last commit.
Questions on content: 
What's the point of starting with a fallacy? Let's go with explaining what they do, not dispelling misconceptions. We don't know if this reader has done that.
What 2 things does git pull do? Seems that should be pull not push, correct? Let's come back to that.